+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.1 (unreleased)
 
+- Added `read_sql` method
 - Added support for symbol keys
 
 ## 0.2.0 (2023-01-14)